agent/0816/onboarding-agent 6/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
httpx==0.27.2
redis==5.0.8
boto3==1.34.162
numpy==1.26.4
pandas==2.2.2
pyarrow==16.1.0
rank-bm25==0.2.2
faiss-cpu==1.8.0
sentence-transformers==3.0.1
torch>=2.2.0
slack_sdk==3.30.0
